fix date-handling bug; stop using inappropriate libs
The definition of ONE_MONTH is apparently some number of days, while the
months being subtracted "across" contain varying numbers of days (leading
to spuriously (depending on the date the program is run) incorrect
results.

Instead, use and perform math on yy.mm results directly from localtime;
in this approach, the number of days is any given month is irrelevant, and
results are consistent.